requirements.txt
Flask==3.1.0
langchain==0.3.14
langchain_community==0.3.14
mistralai==1.2.6
PyPDF2==3.0.1
python-dotenv==1.0.1
